Separated analysis code further
