+++ source/test/sgp_mode_test/SGPModularity.test.cc
@@ -188,43 +188,7 @@ TEST_CASE("GetPModularity", "[sgp]"){
 
 // }
 
-TEST_CASE("GetNecessaryInstructions", "[sgp]"){
-        emp::Random random(10);
-        SymConfigBase config;
-        config.RANDOM_ANCESTOR(false);
-        int world_size = 1;
-        int pop_size = 1;
-        TaskSet test_tasks = LogicTasks;
-        SGPWorld world(random, &config, test_tasks);
-        world.resize(world_size);
-        config.POP_SIZE(pop_size);
-
-        emp::Ptr<SGPHost> test_sample = emp::NewPtr<SGPHost>(&random, &world, &config);
-        std::cout<< world.GetFullPop().size();
-        
-        size_t test_id =0;
-
-       emp::vector<int> program_position_guide = GetNecessaryInstructions(test_sample, test_id, test_tasks); 
-        emp::vector<int> zero_vector = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-        0,0,0,0,0,0,0,0,0,0};
-       REQUIRE(program_position_guide == zero_vector);
-        //right now this returns a vector of  vectors of ints that are all 0
-
-        std::cout<<" -GugGWQ ";
-        
-        
-        
-
-
-}
-
-
-
-
-
-// TEST_CASE("GetReducedProgramRepresentations", "[sgp]"){
+// TEST_CASE("GetNecessaryInstructions", "[sgp]"){
 //         emp::Random random(10);
 //         SymConfigBase config;
 //         config.RANDOM_ANCESTOR(false);
@@ -235,21 +199,57 @@ TEST_CASE("GetNecessaryInstructions", "[sgp]"){
 //         world.resize(world_size);
 //         config.POP_SIZE(pop_size);
 
+//         emp::Ptr<SGPHost> test_sample = emp::NewPtr<SGPHost>(&random, &world, &config);
+//         std::cout<< world.GetFullPop().size();
+        
+//         size_t test_id =0;
+
+//        emp::vector<int> program_position_guide = GetNecessaryInstructions(test_sample, test_id, test_tasks); 
 //         emp::vector<int> zero_vector = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
 //         0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
 //         0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
 //         0,0,0,0,0,0,0,0,0,0};
+//        REQUIRE(program_position_guide == zero_vector);
+        //right now this returns a vector of  vectors of ints that are all 0
 
-//         emp::vector<emp::vector<int>> test_map = {};
-//         emp::Ptr<SGPHost> test_sample = emp::NewPtr<SGPHost>(&random, &world, &config);
-//         std::cout<< world.GetFullPop().size();
-//        test_map = GetReducedProgramRepresentations(test_sample);
-//        emp::vector<emp::vector<int>>zero_map = {};
-        //for(int guide = 0; guide<test_map.size();guide++){
-        //       REQUIRE( test_map ==zero_map);
+//         std::cout<<" -GugGWQ ";
+        
+        
+        
+
+
+// }
+
+
+
+
+
+TEST_CASE("GetReducedProgramRepresentations", "[sgp]"){
+        emp::Random random(10);
+        SymConfigBase config;
+        config.RANDOM_ANCESTOR(false);
+        int world_size = 1;
+        int pop_size = 1;
+        TaskSet test_tasks = LogicTasks;
+        SGPWorld world(random, &config, test_tasks);
+        world.resize(world_size);
+        config.POP_SIZE(pop_size);
+
+        emp::vector<int> zero_vector = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+        0,0,0,0,0,0,0,0,0,0};
+
+        emp::vector<emp::vector<int>> test_map = {};
+        emp::Ptr<SGPHost> test_sample = emp::NewPtr<SGPHost>(&random, &world, &config);
+        std::cout<< world.GetFullPop().size();
+       test_map = GetReducedProgramRepresentations(test_sample);
+       emp::vector<emp::vector<int>>zero_map = {};
+        for(int guide = 0; guide<test_map.size();guide++){
+              REQUIRE( test_map ==zero_map);
              
                //right now this returns a vector of  vectors of ints that are all 0
-        //}
+        }
         
        // GetReducedPrograms(test_sample); // gets segfault
 
@@ -259,7 +259,7 @@ TEST_CASE("GetNecessaryInstructions", "[sgp]"){
 
 
 
-// }
+ }
